Add separate CheckpointDataFile
Remove the Checkpoints from CryptoNoteConfig.h and add a new File for these Informations.
