+++ src/Daemon/Daemon.cpp
@@ -30,6 +30,7 @@
 #include "Common/StringTools.h"
 #include "Common/PathTools.h"
 #include "crypto/hash.h"
+#include "../lib/CryptoNoteCheckpoints.h"
 #include "CryptoNoteCore/CryptoNoteTools.h"
 #include "CryptoNoteCore/Core.h"
 #include "CryptoNoteCore/CoreConfig.h"
